libatalk/vfs: fix handling of short req_count when fetching xattr
This fixes several issue:

1. MAX_REPLY_EXTRA_BYTES was wrong, it is 6, not 8. 6 bytes for the AFP
response bitmap (2) and length (4)

2. We must protect us against bogus MAX_REPLY_EXTRA_BYTES, otherwise
maxreply can underflow or become zero. The latter would be interpreted
as a request to get the xattr size, not the content.

3. If the underylying xattr API reports ERANGE then the xattr is larger
then the requested size, we must return this as AFPERR_PARAM to the
client.

See also commit 66b5b77991679fa8b0b910d159f8592c5414b2b6 and the new
test 432 in our test-suite.

Bug: https://github.com/Netatalk/Netatalk/issues/82

Signed-off-by: Ralph Boehme <slow@samba.org>